+++ src/formats/zeiss_zvi.rs
@@ -1737,7 +1737,7 @@ fn parse_zvi(
         size_c,
         size_t,
         pixel_type,
-        bits_per_pixel: (bytes_per_sample * 8) as u8,
+        bits_per_pixel: (bytes_per_sample * 8) as u16,
         image_count,
         dimension_order,
         is_rgb,
